Dev mat mat evolution

+++ src/include/spin/SpinCollection.h
@@ -34,7 +34,7 @@ public:
     //@}
 
     //@{
-    int           getSpinNum() const {return _spin_list.size();};
+    size_t           getSpinNum() const {return _spin_list.size();};
     vector<cSPIN> getSpinList() const {return _spin_list;};
     vector<cSPIN> getSpinList(const cClusterIndex& clst) const;
     mat getCoordinateMat() const;
